everything is borked

+++ core/src/io/anuke/mindustry/game/Content.java
@@ -6,10 +6,10 @@ import io.anuke.mindustry.type.ContentType;
 
 /** Base class for a content type that is loaded in {@link io.anuke.mindustry.core.ContentLoader}. */
 public abstract class Content{
-    public final byte id;
+    public final short id;
 
     public Content(){
-        this.id = (byte)Vars.content.getBy(getContentType()).size;
+        this.id = (short)Vars.content.getBy(getContentType()).size;
         Vars.content.handleContent(this);
     }
